+++ src/AutoMapper.Collection.Tests/MapCollectionWithEquality.cs
@@ -5,11 +5,11 @@ using AutoMapper.EquivilencyExpression;
 using AutoMapper.Mappers;
 using FluentAssertions;
 
-namespace AutoMapper.Collection.Tests
+namespace AutoMapper.Collection
 {
-    public class MapCollectionWithEquality
+    public class MapCollectionWithEqualityTests
     {
-        public MapCollectionWithEquality()
+        public MapCollectionWithEqualityTests()
         {
             Mapper.Initialize(x =>
             {
@@ -18,31 +18,62 @@ namespace AutoMapper.Collection.Tests
             });
         }
 
-        public void Should_Update_Existing_Item()
+        public void Should_Keep_Existing_List()
         {
-            var dtos = new List<ThingDto>()
+            var dtos = new List<ThingDto>
             {
-                new ThingDto() { ID = 1, Title = "test0" },
-                new ThingDto() { ID = 2, Title = "test2" }
+                new ThingDto { ID = 1, Title = "test0" },
+                new ThingDto { ID = 2, Title = "test2" }
             };
 
-            var items = new List<Thing>()
+            var items = new List<Thing>
             {
-                new Thing() { ID = 1, Title = "test1" },
-                new Thing() { ID = 3, Title = "test3" },
+                new Thing { ID = 1, Title = "test1" },
+                new Thing { ID = 3, Title = "test3" },
+            };
+
+            Mapper.Map(dtos, items).Should().BeSameAs(items);
+        }
+
+        public void Should_Update_Existing_Item()
+        {
+            var dtos = new List<ThingDto>
+            {
+                new ThingDto { ID = 1, Title = "test0" },
+                new ThingDto { ID = 2, Title = "test2" }
+            };
+
+            var items = new List<Thing>
+            {
+                new Thing { ID = 1, Title = "test1" },
+                new Thing { ID = 3, Title = "test3" },
             };
 
             Mapper.Map(dtos, items).Should().HaveElementAt(0, items.First());
         }
+
         public void Should_Work_With_Null_Destination()
         {
-            var dtos = new List<ThingDto>()
+            var dtos = new List<ThingDto>
             {
-                new ThingDto() { ID = 1, Title = "test0" },
-                new ThingDto() { ID = 2, Title = "test2" }
+                new ThingDto { ID = 1, Title = "test0" },
+                new ThingDto { ID = 2, Title = "test2" }
             };
             
             Mapper.Map<List<Thing>>(dtos).Should().HaveSameCount(dtos);
+        }
+
+        public class Thing
+        {
+            public int ID { get; set; }
+            public string Title { get; set; }
+            public override string ToString() { return Title; }
+        }
+
+        public class ThingDto
+        {
+            public int ID { get; set; }
+            public string Title { get; set; }
         }
     }
 }
