--- src/AutoMapper.Collection.Tests/ThingDto.cs
@@ -1,8 +1,0 @@
-namespace AutoMapper.Collection.Tests
-{
-    public class ThingDto
-    {
-        public int ID { get; set; }
-        public string Title { get; set; }
-    }
-}
